feat: multi thread sum of primes first alternative

+++ cpp/include/parallel_math.h
@@ -5,6 +5,9 @@
 #ifndef PARALLEL_PROGRAMMING_PARALLEL_MATH_H
 #define PARALLEL_PROGRAMMING_PARALLEL_MATH_H
 #include <vector>
+#include <thread>
+#include <iostream>
+#include <atomic>
 namespace Samples {
     class ParallelMath {
     private:
@@ -21,7 +24,7 @@ namespace Samples {
 
         static bool is_prime(int number);
 
-        double calculate_sum_primes(const size_t size);
+        unsigned int calculate_sum_primes(const size_t size);
 
         void sum(const std::vector<unsigned int> &a, const std::vector<unsigned int> &b);
     };
